[date-picker] fixed mock for dates to use it in playwright browser tests too
