+++ semcore/date-picker/__tests__/index.test.tsx
@@ -4,6 +4,7 @@ import { expect, test, describe, beforeEach, vi } from '@semcore/testing-utils/v
 import { cleanup, render, fireEvent, act, userEvent } from '@semcore/testing-utils/testing-library';
 import { axe } from '@semcore/testing-utils/axe';
 import { Box, Flex } from '@semcore/flex-box';
+import { mockDate, RealDate } from './utils';
 
 import {
   DatePicker,
@@ -13,24 +14,6 @@ import {
   MonthDateRangeComparator,
 } from '../src';
 
-const RealDate = global.Date;
-
-// https://github.com/facebook/jest/issues/2234#issuecomment-384884729
-function mockDate(isoDate: any) {
-  (global as any).Date = class extends RealDate {
-    constructor(...theArgs: any[]) {
-      super();
-      if (theArgs.length) {
-        return new (RealDate as any)(...theArgs);
-      }
-      return new RealDate(isoDate);
-    }
-
-    static now() {
-      return new RealDate(isoDate).getTime();
-    }
-  };
-}
 describe('DatePicker', () => {
   beforeEach(() => {
     global.Date = RealDate;
